update json/rename files/update styles
