+++ apps/atm-browser/src/assets/styles/_color-vars.scss
@@ -1,3 +1,7 @@
+:root {
+    --bg-d: #252525;
+}
+
 $base-bg-dark: #252525;
 $base-bg-light: #fff;
 $bg-auth-light: #b3b3b3;
