+++ apps/atm-browser/src/assets/styles/common-styles.scss
@@ -1,3 +1,3 @@
-@import './mixins/mixin-dark';
-@import './mixins/mixin-light';
-@import './typography-vars';
+@forward './mixins/mixin-dark';
+@forward './mixins/mixin-light';
+@forward './typography-vars';
