+++ apps/atm-browser/src/assets/styles/mixins/_mixin-dark.scss
@@ -2,4 +2,6 @@
 @mixin darken {
     --base-bg: #{$base-bg-dark};
     --tui-base-01: var(--base-bg);
+    --text-color: #{$base-text-color-dark};
+    --tui-text-01: var(--text-color);
 }
